[major] fix bugs aggreagte service

+++ perpustakaan/peminjaman/src/main/java/com/anla/Peminjaman/VO/Pengembalian.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Data
@@ -13,6 +14,6 @@ public class Pengembalian {
     private Long id;
     private LocalDate tanggalDikembalikan;
     private Integer terlambat;
-    private Double denda;
+    private BigDecimal denda;
     private Long peminjamanId;
 }
